+++ LookupAnything/release-notes.md
@@ -4,7 +4,8 @@
 ## 1.23
 Released 26 November 2019.
 
-* Updated for Stardew Valley 1.4, including new content.
+* Updated for Stardew Valley 1.4, including...
+  * movie theater.
 * Added save format version to save slot lookups.
 * Updated for compatibility with Json Assets 1.3.8.
 * Fixed player luck in multiplayer showing your own luck instead of theirs.